# Conflicts:
#	client/sdl/ECS/KeyboardController.cpp
#	client/sdl/ECS/SpriteComponent.cpp

+++ common/network/messages/server/player/EntityMoveMessage.h
@@ -9,7 +9,7 @@
 
 class EntityMoveMessage : public Message {
 public:
-    EntityMoveMessage(uint8_t id, uint16_t x, uint16_t y);
+    EntityMoveMessage(uint8_t id, int16_t x, int16_t y);
 
     uint8_t opCode() const override;
 
@@ -18,12 +18,12 @@ public:
 
 
     uint8_t getId() const;
-    uint16_t getX()   const ;
-    uint16_t getY()   const;
+    int16_t getX()   const ;
+    int16_t getY()   const;
 private:
     std::uint8_t entityId;
-    float x;
-    float y;
+    uint16_t x;
+    uint16_t y;
 };
